Add AuthContainer
Change Container properties from private to package private.
