+++ skygear_example/src/main/java/io/skygear/skygear_example/LoginActivity.java
@@ -61,7 +61,7 @@ public class LoginActivity extends AppCompatActivity {
         Log.i(LOG_TAG, "doLogin: Signup with email: " + email);
         Log.i(LOG_TAG, "doLogin: Signup with password: " + password);
 
-        this.skygear.loginWithEmail(email, password, new AuthResponseHandler() {
+        this.skygear.auth().loginWithEmail(email, password, new AuthResponseHandler() {
             @Override
             public void onAuthSuccess(User user) {
                 loading.dismiss();
