+++ skygear_example/src/main/java/io/skygear/skygear_example/UserQueryActivity.java
@@ -60,7 +60,7 @@ public class UserQueryActivity extends AppCompatActivity {
                 .setNeutralButton("Dismiss", null)
                 .create();
 
-        this.skygear.getUserByEmail(email, new UserQueryResponseHandler() {
+        this.skygear.auth().getUserByEmail(email, new UserQueryResponseHandler() {
 
             @Override
             public void onQuerySuccess(User[] users) {
